Enforce separate "comment" permission on Sandstorm.
I used to work on Google Docs sharing. One thing I helped implement there was the ability to grant "comment" permission independently of "read" and "write". Originally we only let editors comment, but users made it very clear that they wanted to grant comment permission without write permission. On the other hand, giving comment permission to all readers (as ep_comments currently does) is also not great, because it means public documents end up littered with spam.

Overhauling Etherpad's basic sharing model to include "comment-only" links (separate from the existing read-only links) would probably be a big project. I have not done that here.

However, on Sandstorm, the sharing model is controlled by the platform. The app merely declares what permissions exist, and then Sandstorm provides the UI for users to share various permissions. So, on Sandstorm, we can trivially add a "comment" permission, and I have done so.

This change extends ep_comments with the minimal changes needed to enforce the Sandstorm comment-only permission. I could maintain this as a fork, but the change should be harmless to merge upstream, and this would be a lot easier for me to maintain.