Create a visual gutter on the right for comments to live in.
Previously they overlayed on the document content. But this led to a bug where when no comments were present, the content could flow under the invisible comment sidebar. Clicking in this area would not work, since the click would be eaten by the sidebar. It was also just pretty ugly to do things that way. This is nicer.